# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger for Plask task API",
        "description": "C.R.C 프로젝트 Swagger",
        "license": {
            "name": "MIT",
            "url": "http://10.120.75.224:3000"
        }
    },
    "host": "10.120.75.224:3000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "회원가입",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "회원가입시 필요 데이터 수집",
                        "schema": {
                            "$ref": "#/definitions/RegisterData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Register Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
        "/register/:authNum": {
            "get": {
                "tags": [
                    "Register"
                ],
                "summary": "인증번호 인증 페이지",
                "parameters": [
                    {
                        "in": "body",
                        "name": "query",
                        "description": "Auth Check",
                        "schema": {
                            "$ref": "#/definitions/authNum_Check"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Auth Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
        "/review/check": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "리뷰 목록 확인",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Review Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
        "/review/register": {
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "리뷰 작성",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Review Write",
                        "schema": {
                            "$ref": "#/definitions/Review_Write"
                        }
                    },
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Review Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
        "/review/empathy": {
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "리뷰 공감",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Review empathy",
                        "schema": {
                            "$ref": "#/definitions/Review_empathy"
                        }
                    },
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Review Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
        "/review/reply": {
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "리뷰 응답",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Review reply",
                        "schema": {
                            "$ref": "#/definitions/Review_reply"
                        }
                    },
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Review Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
...
        "/suggest/register": {
            "post": {
                "tags": [
                    "suggest"
                ],
                "summary": "게시판 작성",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "suggest Write",
                        "schema": {
                            "$ref": "#/definitions/Suggest_Write"
                        }
                    },
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                
                "responses": {
                    "201": {
                        "description": "suggest Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
...
        "/notice/register": {
            "post": {
                "tags": [
                    "notice"
                ],
                "summary": "공지사항 등록",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "토큰 확인 및 내용 확인",
                        "schema": {
                            "$ref": "#/definitions/notice_register"
                        }
                    },
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "check Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "RegisterData": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "student_data": {
                    "type": "string"
                }
            }
        },
...
        "authNum_Check": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "authNum": {
                    "type":"string"
                }
            }
        },
        "Review_Write": {
            "properties": {
                "review_star": {
                    "type":"number"
                },
                "content": {
                    "type":"string"
                },
                "nickname": {
                    "type":"string"
                },
                "when": {
                    "type":"string"
                }
            }
        },
        "Review_empathy": {
            "properties": {
                "reviewid": {
                    "type":"integer"
                }
            }
        },
        "Review_reply": {
            "properties": {
                "reply": {
                    "type":"string"
                },
                "reviewid": {
                    "type":"integer"
                }
            }
        },
        "Suggest_Write": {
            "properties": {
                "title": {
                    "type":"string"
                },
                "content": {
                    "type":"string"
                },
                "nickname": {
                    "type":"string"
                },
                "when":{
                    "type":"string"
                }
            }
        },
...
        "notice_register": {
            "properties": {
                "title": {
                    "type":"string"
                },
                "content": {
                    "type":"string"
                }
            }
        }
    }
}